Re-write how OxygenLibrary keywords run subprocesses.
Previously, OxygenLibrary keywords run given commands during test execution
in subprocesses so that return code of the subprocess was always checked to be
0.

Practically, this will not work though; It is fairly common that test tools
report failing test executions with non-zero return codes as well as internal
errors. Thus, we want to enable test designer to decide, if the command should
be allowed to fail, or should it be strictly checked.

Introduce `check_return_code` argument to `utils.run_command_line()`. Re-write
OxygenLibrary keywords to pass this argument from user. With tests.

Enhance keyworddocumentation as well.