Fix utils.run_command_line() to work on Windows and correctly elsewhere
Apparently[1], when you want to pass environment to subprocess.run(), you need
to pass full copy of the entire environment, adding your additions /
modifications. On Windows, this fails spectacularly, but not doing so is not
correct on *nixes either -- this seems to fix what was earlier thought to be
normal limitation of not being able to run subprocesses inside other
subprocesses; you can, if you do it right!

In this commit, refactor utils.run_command_line() to always pass full
environment. Also tests.

[1] https://www.scivision.dev/python-calling-python-subprocess/