Fix time problems on Windows 10
Due to Windows 10[1], Robot Interfaces did produce timestamps that were
before epoch and thus failed when Robot tried to convert them into
results.

Now, if we try to put a timestamp that is before epoch, it is just reset
to epoch.

Also, RF3 interface did not have parity with the RF4 interface -- now it
does

[1] https://stackoverflow.com/a/58203399